requirements.txt
cryptography
pycryptodome
pydantic
pyjwt
requests-cache
